requirements.txt
Flask>=3.0
markdown>=3.6
python-frontmatter>=1.1
Pygments>=2.17
